backend/pyproject.toml
[project]
name = "loglens-api"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "python-dotenv==1.0.1",
    "pydantic==2.10.3",
    "httpx==0.28.1",
    "slowapi==0.1.9",
    "structlog==24.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pytest-asyncio]
mode = "auto"
